fix some bug on paco -t 1

+++ includes/ft_printf_bonus.h
@@ -6,7 +6,7 @@
 /*   By: llevasse <llevasse@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/04 16:39:11 by llevasse          #+#    #+#             */
-/*   Updated: 2023/02/09 15:13:26 by llevasse         ###   ########.fr       */
+/*   Updated: 2023/02/09 15:47:13 by llevasse         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,7 +56,7 @@ void	print_width_prec_s(char *str_to_print, int prec, int *sum);
 int		predict_len(const char *str, va_list args);
 int		get_size_base_u(unsigned long long n, const char *base);
 int		get_size_base(int n, const char *base);
-int		get_int_len(long long n);
+int		get_int_len(int n);
 int		get_unsigned_len(unsigned int n);
 int		print_odd(const char *str, va_list args, int *sum, int *padding);
