+++ srcs/var_bonus/print_field_width.c
@@ -6,7 +6,7 @@
 /*   By: llevasse <llevasse@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/28 10:28:18 by llevasse          #+#    #+#             */
-/*   Updated: 2023/02/09 14:42:00 by llevasse         ###   ########.fr       */
+/*   Updated: 2023/02/09 15:54:25 by llevasse         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,7 +61,7 @@ int	is_neg(const char *str, va_list args)
 	va_copy(args_cp, args);
 	if (ft_is_in_str("dixXu", get_spec(str)))
 	{
-		int_to_print = va_arg(args_cp, long long);
+		int_to_print = va_arg(args_cp, int);
 		if (get_para(str) == '0')
 		{
 			if (int_to_print < 0)
